Modify wrapper and shell functions to handle non-CGI use better
The setuid wrapper and shell utility library now omit the Content-Type
header if no QUERY_STRING environment variable is set (empty QS still
counts as set).

This means that when invoked from the command line rather than a CGI
handler we get JSON output which can be passed directly to jq et al.

In addition the setuid wrapper now sanitises its own command line args
and passes then to the worker scripts instead of the QUERY_STRING
so invoking the scripts from the command line is easier:

  QUERY_STRING=foo=bar&blerg=weasel

is equivalent to

  /path/to/wrapper/hardlink foo=bar blerg=weasel