Use the alternate reset image manifest from the atomupd server
The static server is now capable of telling us which image to
install if we (for whatever reason) don't want the latest image
for our current variant.

Typically we need to do this if we're on the latest image since
we cannot currently install the same image in both OS slots.