os-status: Use the new --penultimate-update option
The meta URL structure changed since the introduction of the branch
concept.

Instead of fixing every time the URL, we should stop to open code it and
start to use the `--penultimate-update` option.
This will ensure that we are always looking for updates in the correct
place.

The `--penultimate-update` option has been introduced since
steamos-atomupd 0.20240207.1.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>