steamos-reset-tool: reinstate default behaviour of a full reset
When the option of resetting either the user data or the OS separately
was introduced steamos-reset-tool started requiring an explicit reset
mode command line argument.

This broke the reset-from-steam-ui flow which invokes the reset tool
without any addition arguments: This commit restores the old no-args
behaviour of starting a full reset.