Ensure that resetting OS and user data waits for both to complete
Currently there is a bug if the user (e.g. Steam) calls
`factory-reset --reset-os --reset-user-data`; because the user data
reset runs in parallel to the OS reset, it will report its success
first, which then causes `factory-reset-tool` to exit, and for Steam to
think that the reset is finished, however the OS reset is still
occurring.  In the case where the OS reset takes a long time to
complete, this means that most likely the OS reset does not complete by
the time the user hits the reboot button.

Fix this by (1) serializing these two reset operations (this has
negligible time impact becuase the user data reset takes almost no time
at all) and (2) ensuring that the first operation does not report
success, only reporting success if both operations succeed.  If the OS
operation fails, we will immediately report failure, which will cause
the parent `steamos-reset-tool` process to fail out.